tests/requirements.txt
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
iniconfig==1.1.1
jsonschema==3.2.0
mysql-connector-python==8.0.27
packaging==21.2
paramiko==2.8.0
pluggy==1.0.0
protobuf==3.19.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-docker==0.10.3
pytest-dotenv==0.5.2
python-dotenv==0.19.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
texttable==1.6.4
toml==0.10.2
urllib3==1.26.7
websocket-client==0.59.0
